Conflicts:
	.github/workflows/submit.yml
	make/Docs.gmk
	make/test/BuildMicrobenchmark.gmk
	src/java.base/share/classes/java/lang/invoke/DirectMethodHandle.java
	src/java.base/share/classes/java/lang/reflect/ProxyGenerator.java
	src/java.base/share/classes/jdk/internal/reflect/UnsafeObjectFieldAccessorImpl.java
	src/java.base/share/classes/jdk/internal/reflect/UnsafeQualifiedObjectFieldAccessorImpl.java
	src/java.base/share/classes/jdk/internal/reflect/UnsafeQualifiedStaticObjectFieldAccessorImpl.java
	src/java.base/share/classes/jdk/internal/reflect/UnsafeStaticObjectFieldAccessorImpl.java
	src/java.base/share/native/include/jni.h
	src/jdk.compiler/share/classes/com/sun/tools/javac/code/Source.java
	src/jdk.compiler/share/classes/com/sun/tools/javac/jvm/ClassWriter.java
	src/jdk.compiler/share/classes/com/sun/tools/javac/resources/compiler.properties
	test/hotspot/jtreg/runtime/ErrorHandling/MachCodeFramesInErrorFile.java
	test/jdk/TEST.groups

+++ src/java.base/share/classes/jdk/internal/org/objectweb/asm/Frame.java
@@ -613,11 +613,11 @@ class Frame {
       */
     private void pop(final int elements) {
         if (outputStackTop >= elements) {
-            outputStackTop -= elements;
+            outputStackTop -= (short) elements;
         } else {
             // If the number of elements to be popped is greater than the number of elements in the output
             // stack, clear it, and pop the remaining elements from the input stack.
-            outputStackStart -= elements - outputStackTop;
+            outputStackStart -= (short) (elements - outputStackTop);
             outputStackTop = 0;
         }
     }
@@ -1505,4 +1505,3 @@ class Frame {
         }
     }
 }
-
